# openapi.yaml
openapi: '3.0.1'
info:
  version: 1.0.0
  title: Product-store
  description: A simple Product Store
servers: 
  - url: https://localhost:3000
paths:
...
  /customers/signup:
    post:
      description: signs up a new user in the database
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: John
                surname:
                  type: string
                  example: Doe
                contact:
                  type: number
                  example: 777999888
                username:
                  type: string
                  example: john_doe
                password:
                  type: string
                  example: pass1234
  
      responses:
        201: 
          description: Customer was added to the database
          content:
            application/json:
              schema:
                type: object
                example: 
                  message: Customer added successfully
        409:
          description: Conflict - customer already exists
          content:
            application/json:
              schema:
                type: object
                example: 
                  error: Customer already exists
  /customers/login:
    post:
      description: signs up a new user in the database
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: John
                surname:
                  type: string
                  example: Doe
                contact:
                  type: number
                  example: 777999888
                username:
                  type: string
                  example: john_doe
                password:
                  type: string
                  example: pass1234
  
      responses:
        200: 
          description: Customer successfully logged in
          content:
            application/json:
              schema:
                type: object
                example: 
                  username: john
                  token: big_token_here
        401:
          description: Incorrect credentials
          content:
            application/json:
              schema:
                type: object
                example: 
                  error: Incorrect password
        404:
          description: Customer not found
          content:
            application/json:
              schema:
                type: object
                example: 
                  error: Customer not found